Refactor SecurityEvent model and add tests
