+++ src/SecuNik.Core/Models/ParsingModels.cs
@@ -1,27 +1,6 @@
 using System.Collections.Generic;
 namespace SecuNik.Core.Models
 {
-    /// <summary>
-    /// Request model for file analysis
-    /// </summary>
-    public class AnalysisRequest
-    {
-        public string FilePath { get; set; } = string.Empty;
-        public string? OriginalFileName { get; set; }
-        public AnalysisOptions Options { get; set; } = new();
-    }
-
-    /// <summary>
-    /// Analysis configuration options
-    /// </summary>
-    public class AnalysisOptions
-    {
-        public bool EnableAIAnalysis { get; set; } = true;
-        public bool GenerateExecutiveReport { get; set; } = true;
-        public bool IncludeTimeline { get; set; } = true;
-        public int MaxSecurityEvents { get; set; } = 10000;
-        public List<string> FocusKeywords { get; set; } = new();
-    }
 
     /// <summary>
     /// Supported file types for parsing
