--- tests/SecuNik.Core.Tests/UnitTest1.cs
@@ -1,10 +1,0 @@
-namespace SecuNik.Core.Tests;
-
-public class UnitTest1
-{
-    [Fact]
-    public void Test1()
-    {
-
-    }
-}
